Implement the function to add a user to db.json
