+++ src/html/login.html
@@ -121,7 +121,7 @@
       </div>
     </div>
   </footer>
-  <script type="module" src="../js/login_json.js"></script>
+  <script type="module" src="../js/login.js"></script>
 </body>
 
 </html>
