+++ src/js/login.js
@@ -1,27 +1,38 @@
-// Simulação de um banco de dados com JSON
-const users = [
-  { username: "usuario1", password: "senha123" },
-  { username: "usuario2", password: "senha456" }
-];
+import axios from 'axios';
 
-// Função para validar o login
-function validateLogin() {
-  const username = document.getElementById("loginName").value;
-  const password = document.getElementById("loginPassword").value;
-  const messageElement = document.getElementById("message");
+async function registerLoginCredential() {
+    
+    const login = document.getElementById("txtLogin").value;
+    const password = document.getElementById("txtSenha").value;
+    const name = document.getElementById("txtNome").value;
+    const email = document.getElementById("txtEmail").value;
 
-  // Verifica se as credenciais estão corretas
-  const user = users.find(user => user.username === username && user.password === password);
+    const inputCredential = {
+        login,
+        password,
+        name,
+        email
+    };
 
-  if (user) {
-      messageElement.textContent = "Login bem-sucedido!";
-      messageElement.style.color = "green";
-      // Aqui você pode redirecionar para outra página ou realizar outra ação
-  } else {
-      messageElement.textContent = "Credenciais inválidas. Tente novamente.";
-      messageElement.style.color = "red";
-  }
+    await axios.post(`http://localhost:3000/login_credentials`, inputCredential);
+    document.getElementById('registerForm').reset();
 }
 
-// Adiciona o evento de clique ao botão de login
-document.getElementById("loginButton").addEventListener("click", validateLogin);
+async function checkIfLoginExists(inputCredential) {
+    const { data: credential_list } = await axios.get("http://localhost:3000/login_credentials");
+
+    foundLogin = credential_list.find(credential => credential.username === inputCredential.username);
+
+    if (foundLogin && foundLogin.password === inputCredential.password) {
+        // login successful
+        console.log("LOGIN SUCESSFULL!!!");
+    } else {
+        // login failed
+        console.log("Please check you login and password!");
+    }
+
+}
+
+// include functions on global scope of the window
+window.registerLoginCredential = registerLoginCredential;
+window.checkIfLoginExists = checkIfLoginExists;
